pyproject.toml
[project]
name = "memory-platform"
version = "0.1.0"
description = "AI Memory Platform based on mem0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    # mem0 fork — 外部依赖提升为直接依赖
    "qdrant-client>=1.9.1",
    "openai>=1.90.0",
    "posthog>=3.5.0",
    "pytz>=2024.1",
    "sqlalchemy>=2.0.31",
    "protobuf>=5.29.6,<7.0.0",
    # mem0 传递依赖
    "httpx>=0.28",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "anthropic>=0.89.0",
    "tcvectordb>=1.3.0",
    # MySQL
    "pymysql>=1.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/memory_platform", "src/mem0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100
